refactor(cli): extract analyzer route breakdown helpers
